Updated README in all sample apps (#255)
Updated the README file in all sample applications to include an
informational message at the top that before the various commands are
run, they will need to be in the correct folder.

Often folks forget to be in the right application folder before
executing the commands. This causes the commands to fail. The addition
of the instruction above will help to ensure that the user is in the
right folder and then the commands are run.

---------

Co-authored-by: Polong Lin <polong-lin@users.noreply.github.com>