[Gemini function calling] Simplify prompt in chat session to avoid multiple function calls (#368)
Gemini will sometimes now return two or more subsequent function calls
after a function response so that it can gather more information in a
single conversation turn with the end user. This broke the simple
example in the Gemini function calling notebook since (for the sake of
simplicity) it assumes one function call per conversation turn.

```python
>>> response = chat.send_message(
>>>    Part.from_function_response(
>>>        name="place_order",
>>>        response={
>>>            "content": api_response,
>>>        },
>>>    ),
>>>)
>>>
>>>response.candidates[0].content.parts[0]

InvalidArgument: 400 Please ensure that function response turn comes immediately after a function call turn.
```

This PR simplifies one of the prompts to keep the conversation flow in
the notebook on track and error-free for this walkthrough.