fix: when using seed parameter, watermarking must be disabled. (#780)
# Description

Running the notebook in Colab Enterprise. Ran across an issue where when
using the seed parameter the call would fail if watermarking was not set
to false (it is true by default). The specific error returned was:

details = "Image generation failed with the following error: Seed is not
supported when watermark is enabled."

Updated to include code for any seed parameter calls to turn off
watermarking. You may want to consider reordering so the watermarking
section occurs before the seed section.

---------

Co-authored-by: Holt Skinner <holtskinner@google.com>