docs(gemini-streamlit-cloudrun): Remove separate build step (#811)
# Description

This change removes the build step for the Cloud Run deployment
instructions. Instead, the Cloud Run deploy uses a source deploy, where
Cloud Run's CLI operation manages the steps to upload and build the
container image for use.

This simplifies the instructions and accelerates folks to using the
sample app more quickly.