ci: minimize pull_request_target usage (#887)
# Description

This change removes `pull_request_target` from spellcheck, and comments
on it for auto-labeling. A result of this change removes commenting from
the spellcheck action operation.

If you review the PR timeline, you will see messages from two versions
of the spellcheck workflow. This is because the version committed to
main and the version modified in this branch were both operating on the
code.

---------

Co-authored-by: Owl Bot <gcf-owl-bot[bot]@users.noreply.github.com>
Co-authored-by: Holt Skinner <13262395+holtskinner@users.noreply.github.com>