fix: fix error url concat in evaluate_gemini_with_autosxs.ipynb
Using `os.path.join` to concatenate GCS URL is a misuse. The function is os dependent and it returns "\" on a Windows OS, which causes invalid GCS URL like "gs://bucket\path".

A more robust way to concat any URL is to use `urllib.parse.urljoin`. However, given how `BUCKET_URI` and `display_name` are set in this notebook, we can simply use f-string to concate the GCS URI.