fix: Change person_generation filter from allow_all to allow_adult in imagen3_image_generation notebook (#1019)
# Description
The imagen3 model requires a person_generation filter. This is currently
set to "allow_all" in the notebook which will not work unless one has
specifically been given access to this. The most common filter that will
work is "allow_adult". Hence this change sets the filter values to
"allow_adult".

Fixes #1018 🦕

---------

Co-authored-by: Holt Skinner <13262395+holtskinner@users.noreply.github.com>
Co-authored-by: Owl Bot <gcf-owl-bot[bot]@users.noreply.github.com>
Co-authored-by: Holt Skinner <holtskinner@google.com>