feat: Improve error catching in the SQL Talk app (Gemini Function Calling) (#1136)
# Description

This PR adds improved error catching to the SQL Talk app. Currently, if
an error is encountered due to a malformed generated SQL query or an
error executing the tools / functions, then a full stack trace will
appear in the app. With this PR, errors are caught at the SQL execution
level and top-level application and rendered in the app (and persisted
in the message history) without a full stack trace.

Special thanks to @mona19 for suggesting this change and providing a
sample implementation! 🙏

How friendlier errors will appear now instead of a full stack trace:

---

![Screenshot 2024-09-18 at 6 21
37 PM](https://github.com/user-attachments/assets/44c45e96-095e-4373-ab23-a6fde871b6e0)

---

![Screenshot 2024-09-18 at 6 22
11 PM](https://github.com/user-attachments/assets/b47cba62-6ee2-4ddf-8832-9f4e4643d0df)

---------

Co-authored-by: Owl Bot <gcf-owl-bot[bot]@users.noreply.github.com>