fix: Change exclude_patterns to a set to avoid TypeError (#1898)
This pull request addresses a TypeError that occurs when passing
exclude_patterns argument as a list to the ingest function from the
gitingest library. The error message "TypeError: unhashable type:
'list'" indicates that the library's internal logic requires
"exclude_patterns" to be a hashable type, such as a set.

This PR changes the "exclude_patterns" variable from a list to a set
using curly braces {}. This ensures that the exclude_patterns argument
is of the correct type, preventing the TypeError and ensuring the ingest
function works as intended.

This change was made in response to encountering this error while
working with the code in colab, believing this change is a necessary
improvement to prevent similar issues for other users.