fix: issues with pip command in multimodal dataset intro
Currently the pip command prints
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed
```
and restarting the session afterwards leads to a session crash.

This change fixes both of these issues.
I also added a note that a session restart is needed after the pip
command.